+++ term/password.go
@@ -2,17 +2,17 @@ package term
 
 import (
 	"fmt"
+	"syscall"
 
-	"github.com/howeyc/gopass"
-
+	"golang.org/x/term"
 	"gopkg.in/errgo.v1"
 )
 
 func Password(prompt string) (string, error) {
-	fmt.Printf(prompt)
-	bytePassword, err := gopass.GetPasswd()
+	fmt.Print(prompt)
+	bytePassword, err := term.ReadPassword(syscall.Stdin)
 	if err != nil {
-		return "", errgo.Mask(err)
+		return "", errgo.Notef(err, "fail to read the password on stdin")
 	} else {
 		return string(bytePassword), nil
 	}